Rename and organize files and folders to respect the MVC pattern design
