+++ cafeteria_alina/controller/__init__.py
@@ -1,0 +1,3 @@
+## Connects the view (templates) with the model, and process data
+## between them.
+## Pass¨
