+++ cafeteria_alina/controller/auth.py
@@ -1,0 +1,13 @@
+import functools
+
+from os import error
+
+from flask import (render_template, Blueprint, flash, g, redirect, request, session, url_for)
+
+from werkzeug.security import check_password_hash, generate_password_hash
+
+from model import db
+
+
+
+
