+++ cafeteria_alina/model/__init__.py
@@ -1,3 +1,3 @@
 ## The father of all Entities. This is only to make all files as children
-
+## These objects are the relation tables.
 ## Pass¨
