+++ cafeteria_alina/model/repository/__init__.py
@@ -1,0 +1,1 @@
+## The queries used in the app
