+++ cafeteria_alina/model/repository/modelo_producto.py
@@ -1,8 +1,8 @@
 # Import the entity to read as Entity object
-from entity.producto import Producto
+from model.producto import Producto
 
 # Import the database
-from entity.__init__ import db
+from model import db
 
 
 def get_producto(nombre):
@@ -12,4 +12,4 @@ def get_producto(nombre):
 
 def get_producto_id(id_producto):
     '''Regresa un producto más en especifico dado su id (gtin)'''
-    return Producto.query.filter(Producto.id_producto = id_producto)
+    return Producto.query.filter(Producto.id_producto == id_producto)
